refactoring: make clients public
