+++ clients/datasource/http_auth.go
